+++ spec/ltsr.spec.js
@@ -1,33 +1,48 @@
 const expect = require('chai')
   .expect;
 
-const { renderer } = require('../ltsr');
+const LTSR = require('../ltsr');
 const { join: joinPath } = require('path');
 
-describe('ltsr', () => {
-  set('render', () => renderer(joinPath(__dirname, 'mock')));
+describe('LTSR', () => {
+  set('ltsr', () => new LTSR(joinPath(__dirname, 'mock')));
+
+  set('keepWhitespace', false);
+  set('args', () => ({ keepWhitespace }));
 
   sharedExamples('bad render avoidance', () => {
     context('with a path out of root', () => {
       set('path', '../out/of/root');
 
-      it('fails', () => expect(() => rendered).to.throw(Error, /outside of root/));
+      it('fails', () => expect(() => render).to.throw(Error, /outside of root/));
     });
 
     context('with a missing template', () => {
       set('path', 'nothing');
 
-      it('fails', () => expect(() => rendered).to.throw(Error, /ENOENT/));
+      it('fails', () => expect(() => render).to.throw(Error, /ENOENT/));
     });
   });
 
-  describe('render', () => {
-    subject('rendered', () => render(path, args));
+  sharedExamples('whitespace appreciator', () => {
+    describe('whitespace', () => {
+      set('path', 'simple');
 
-    set('args', {});
+      it('does not end with whitespace', () => expect(render).not.to.match(/\s$/g));
+
+      context('when keeping whitespace', () => {
+        set('keepWhitespace', true);
+
+        it('ends with whitespace', () => expect(render).to.match(/\n$/g));
+      });
+    });
+  });
+
+  describe('.render', () => {
+    subject('render', () => ltsr.render(path, args));
 
     itBehavesLike('bad render avoidance');
-
+    itBehavesLike('whitespace appreciator');
 
     context('rendering one level deep', () => {
       set('path', 'index');
@@ -41,18 +56,18 @@ describe('ltsr', () => {
 
         set('args', { locals: new Thing() });
 
-        it('succeeds', () => expect(rendered).to.eq('Bbc'));
+        it('succeeds', () => expect(render).to.eq('Bbc'));
       });
 
       context('when a variable is missing', () => {
         set('args', { locals: { key: 1 } });
-        it('fails', () => expect(() => rendered).to.throw(ReferenceError, /Render failed/));
+        it('fails', () => expect(() => render).to.throw(/Render failed/));
       });
 
       context('with required args present', () => {
         set('args', { locals: { key: 1, val: 2, constant: 3 } });
 
-        it('succeeds', () => expect(rendered).to.eq('123'));
+        it('succeeds', () => expect(render).to.eq('123'));
       });
 
       context('with an Array', () => {
@@ -62,7 +77,7 @@ describe('ltsr', () => {
           keyName: 'key',
           valueName: 'val'
         });
-        it('succeeds', () => expect(rendered).to.eq('0A31B32C3'));
+        it('succeeds', () => expect(render).to.eq('0A31B32C3'));
       });
 
       context('with a Set', () => {
@@ -72,7 +87,7 @@ describe('ltsr', () => {
           keyName: 'key',
           valueName: 'val'
         });
-        it('succeeds', () => expect(rendered).to.eq('0a31b32c3'));
+        it('succeeds', () => expect(render).to.eq('0a31b32c3'));
       });
 
       context('with an object', () => {
@@ -82,7 +97,7 @@ describe('ltsr', () => {
           valueName: 'val'
         });
 
-        it('succeeds', () => expect(rendered).to.eq('a13b23'));
+        it('succeeds', () => expect(render).to.eq('a13b23'));
       });
 
       context('with a map', () => {
@@ -92,7 +107,7 @@ describe('ltsr', () => {
           valueName: 'val'
         });
 
-        it('succeeds', () => expect(rendered).to.eq('A13B23'));
+        it('succeeds', () => expect(render).to.eq('A13B23'));
       });
     });
 
@@ -100,12 +115,12 @@ describe('ltsr', () => {
       set('path', 'outer');
       set('args', { locals: { external: { key: 1, val: 2, constant: 3 } } });
 
-      it('uses the same root', () => expect(rendered).to.eq('z123y'));
+      it('uses the same root', () => expect(render).to.eq('z123y'));
 
       context('when render is set', () => {
         set('args', { locals: { render: null, external: { key: 1, val: 2, constant: 3 } } });
 
-        it('does not overwrite', () => expect(rendered).to.eq('z123y'));
+        it('does not overwrite', () => expect(render).to.eq('z123y'));
       });
 
       context('depth requiring methods', () => {
@@ -113,30 +128,46 @@ describe('ltsr', () => {
           method(value) { return '1' + value; }
         }
         set('args', { locals: new Thing() });
-
         set('path', 'sendingFunctions');
 
-        it('succeeds', () => expect(rendered).to.eq('A1xZ'));
+        it('succeeds', () => expect(render).to.eq('A1xZ'));
+      });
+
+      context('when a render error occurs at depth', () => {
+        set('path', 'rendersBadInner');
+
+        it('presents a useful error', () => {
+          try {
+            render;
+            /* c8 ignore next */
+            expect(1).to.eq(2); // deliberate failure - shouldn't be reached
+          } catch (error) {
+            expect(error.stack).to.include("mock/badInner.lt:3");
+            expect(error.stack).to.include("mock/rendersBadInner.lt:");
+            expect(error.stack.indexOf('/badInner')).to.be.below(error.stack.indexOf('/rendersBadInner'));
+          }
+        });
       });
     });
   });
 
-  describe('render.raw', () => {
-    subject('rendered', () => render.raw(path));
+  describe('.raw', () => {
+    subject('render', () => ltsr.raw(path, keepWhitespace));
 
     itBehavesLike('bad render avoidance');
+    itBehavesLike('whitespace appreciator');
 
     context('with a valid path', () => {
       set('path', 'index');
 
-      it('succeeds', () => expect(rendered).to.eq('${key}${val}${constant}'));
+      it('succeeds', () => expect(render).to.eq('${key}${val}${constant}'));
     });
 
     context('depth raw called', () => {
-      set('rendered', () => render(path));
+      set('render', () => ltsr.render(path));
       set('path', 'rawOuter');
 
-      it('does not try to interpolate at depth', () => expect(rendered).to.eq('z${key}${val}${constant}y'));
+      it('does not try to interpolate at depth', () => expect(render).to.eq('z${key}${val}${constant}y'));
     });
   });
 });